refactor to self contained components
